+++ src/effects/Biquad.h
@@ -35,7 +35,9 @@ struct Biquad
 
    inline float ProcessOne(float fIn)
    {
-      float fOut = fIn * fNumerCoeffs[B0] +
+      // Biquad must use double for all calculations. Otherwise some
+      // filters may have catastrophic rounding errors!
+      double fOut = double(fIn) * fNumerCoeffs[B0] +
             fPrevIn * fNumerCoeffs[B1] +
             fPrevPrevIn * fNumerCoeffs[B2] -
             fPrevOut * fDenomCoeffs[A1] -
@@ -47,12 +49,12 @@ struct Biquad
       return fOut;
    }
 
-   float fNumerCoeffs[3]; // B0 B1 B2
-   float fDenomCoeffs[2]; // A1 A2, A0 == 1.0
-   float fPrevIn;
-   float fPrevPrevIn;
-   float fPrevOut;
-   float fPrevPrevOut;
+   double fNumerCoeffs[3]; // B0 B1 B2
+   double fDenomCoeffs[2]; // A1 A2, A0 == 1.0
+   double fPrevIn;
+   double fPrevPrevIn;
+   double fPrevOut;
+   double fPrevPrevOut;
 
    enum kSubTypes
    {
@@ -65,9 +67,10 @@ struct Biquad
    static ArrayOf<Biquad> CalcChebyshevType1Filter(int order, double fn, double fc, double ripple, int type);
    static ArrayOf<Biquad> CalcChebyshevType2Filter(int order, double fn, double fc, double ripple, int type);
 
-   static void ComplexDiv (float fNumerR, float fNumerI, float fDenomR, float fDenomI, float* pfQuotientR, float* pfQuotientI);
-   static bool BilinTransform (float fSX, float fSY, float* pfZX, float* pfZY);
-   static float Calc2D_DistSqr (float fX1, float fY1, float fX2, float fY2);
+   static void ComplexDiv (double fNumerR, double fNumerI, double fDenomR, double fDenomI,
+                           double* pfQuotientR, double* pfQuotientI);
+   static bool BilinTransform (double fSX, double fSY, double* pfZX, double* pfZY);
+   static float Calc2D_DistSqr (double fX1, double fY1, double fX2, double fY2);
 
    static const double s_fChebyCoeffs[MAX_Order][MAX_Order + 1];
    static double ChebyPoly(int Order, double NormFreq);
